+++ app/config/dependencies_config.py
@@ -1,0 +1,12 @@
+from config.database_config import Base, SessionLocal, engine
+
+
+Base.metadata.create_all(bind=engine)
+
+
+def get_db():
+    db = SessionLocal()
+    try:
+        yield db
+    finally:
+        db.close()
