--- app/dto/credit_card_dto.py
@@ -1,8 +1,0 @@
-from pydantic import BaseModel
-
-
-class CreditCardDto(BaseModel):
-    card_number: str
-    card_holder: str
-    expiration_date: str
-    cvv: str
